Restructured project folders and modules

+++ FASTAflow/flask_app.py
@@ -4,19 +4,19 @@ This module initializes and configures the FASTAflow Flask application.
 It defines a factory function that creates and configures the Flask application,
 initializes the database connection, and sets up necessary components like blueprints.
 """
-
-__author__ = 'Sam Nelen'
-__version__ = '2025.01.16'
-
 import os
-
+from pathlib import Path
 from flask import Flask
+from flask_migrate import Migrate
+from werkzeug.middleware.proxy_fix import ProxyFix
 
-from modules.models import db
-from modules import pages
+from FASTAflow import __version__
+from FASTAflow.modules.models import db
+from FASTAflow.modules import pages
+from FASTAflow.config import DevelopmentConfig, ProductionConfig
 
 
-def create_app():
+def create_app(config_object=None):
     """Creates and configures the Flask application.
 
     Creates a new Flask application instance, configures the database connection,
@@ -28,24 +28,37 @@ def create_app():
     Raises:
         OSError: If database directory creation fails.
     """
-    webapp = Flask(__name__, static_folder='static')
+    app = Flask(__name__, static_folder='static')
 
-    # Create a 'database' directory in your project root
-    db_folder = os.path.join(os.path.dirname(os.path.dirname(__file__)), 'database')
-    os.makedirs(db_folder, exist_ok=True)
+    # Configure the application which defaults to DevelopmentConfig
+    if config_object is None:
+        config_object = ProductionConfig if os.getenv('FLASK_ENV') == 'production' else DevelopmentConfig
+    app.config.from_object(config_object)
 
-    db_path = os.path.join(db_folder, 'fastaflow.db')
-    webapp.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///{db_path}'
-    webapp.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+    # Make sure the required directories exist
+    Path(app.config['DATABASE_DIR']).mkdir(parents=True, exist_ok=True)
+    Path(app.config['UPLOAD_DIR']).mkdir(parents=True, exist_ok=True)
 
-    webapp.secret_key = 'z\x16_\x0f\xe2N\xdd\x83^\x07!<'
+    # Initialize extensions
+    db.init_app(app)
+    Migrate(app, db)
 
-    db.init_app(webapp)
+    # Register blueprints
+    app.register_blueprint(pages.bp)
 
-    with webapp.app_context():
-        db.drop_all()
+    # Configure middleware
+    app.wsgi_app = ProxyFix(app.wsgi_app, x_proto=1, x_host=1)  # Handle proxy headers
+
+    # Add version info into template context
+    @app.context_processor
+    def inject_version():
+        return dict(version=__version__)
+
+    # Initialize database
+    with app.app_context():
+        if app.config['DEBUG']:
+            db.drop_all()  # Only in development
         db.create_all()
-        print(f'database created at {db_path}')
+        app.logger.info(f"Database initialized at {app.config['SQLALCHEMY_DATABASE_URI']}")
 
-    webapp.register_blueprint(pages.bp)
-    return webapp
+    return app
